feat(iotdevice): 使用新的Nacos客户端重构服务配置初始化
- 将原有的全局Nacos初始化函数替换为新的Nacos客户端实例
- 添加Nacos客户端创建方法NewNacosClient
- 修改服务上下文中的Nacos配置和注册逻辑
- 注释掉旧的初始化方式，采用新客户端模式

feat(iotws): 更新设备端口类型并集成设备客户端

- 将DevicePort字段类型从int改为uint64
- 添加DeviceCli字段用于设备通信客户端
- 集成新的设备客户端创建逻辑
- 移除未使用的fmt导入，添加必要的net/http和devicecli导入

feat(devicecli): 新增设备客户端模块

- 创建新的devicecli包提供设备通信客户端
- 实现NewDeviceCli函数支持nacos模式下的动态服务器发现
- 提供HTTP客户端创建功能，支持代理配置

refactor(nacos): 重构Nacos客户端实现为结构体方法

- 将全局函数转换为nacosClient结构体的方法
- 分离配置监听和实例注册功能为独立方法
- 添加GetSeverCli方法用于获取健康的服务实例
- 修复模型判断条件确保正确处理本地模式

+++ app/iotws/cmd/internal/config/config.go
@@ -14,6 +14,6 @@ type Config struct {
 	CacheRedis redis.RedisConf
 
 	DeviceServer string
-	DevicePort   int
+	DevicePort   uint64
 	DviceHost    string
 }
